+++ cmd/vm-network-policy/main.go
@@ -4,12 +4,13 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/harvester/harvester/pkg/firewall/controllers"
 	"github.com/sirupsen/logrus"
+
+	"github.com/harvester/harvester/pkg/vmnetworkpolicy/controllers"
 )
 
 func main() {
-	logrus.Info("Starting firewall controller")
+	logrus.Info("Starting vmnetworkpolicy controller")
 	err := controllers.NewVMNetworkPolicyHandler(os.Getenv("KUBECONFIG"))
 	if err != nil {
 		panic(fmt.Errorf("failed to find kubeconfig: %v", err))
